prime

+++ number.py
@@ -95,7 +95,7 @@ class Number:
         Returns:
             bool: true if number is prime otherwise false.
         ''' 
-        for i in range(2,math.sqrt(self.number)+1):
+        for i in range(2,int(math.sqrt(self.number))+1):
             if self.number %i:
                 return False
         return True
